First stab at recovering the element id from the file location to identify issues by id rather than expression.
This works, but issues remain. The XML parser only finds the id of `element` elements in StructureDefinitions, whereas the JSON parser finds the id of any element. Both may be valid approaches; this feature is first and foremost aimed at StructureDefinitions, where the id by convention is more useful than the experession.
An alternative might be to use `path` and `slicename` explicitly instead of `id`, which is only built up by convention from these elements. In this case, the feature will be explicitly aimed at StructureDefinitions. On the other hand, it could be a nice feature to use `id` for other use cases than StructureDefinitions.

Also, some efficiency improvements need to be made.