Widen the id matching mechanism so that it works on any resource, nut just StructureDefinitions. The key in the known issues file can now simply be an id or a path." I think this ultimately is a more flexible mechanism and retrieving id's turns out to be very fast, so there's little reason not to do it.
Two issues remain however. This does not work on snapshotted StructureDefinitions, the Validator seems to report the wrong line and column numbers for these (in addition, id's turn out not to be unique when you have a snapshot and a differential, but that's irrelevant here). Second, there's a bug in the Validator where column numbers are not reported correctly for constraints.